Title: 합과 차의 나눗셈, Time: 142ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240102/합과 차의 나눗셈/divide-of-sum-and-sub.java
@@ -3,8 +3,8 @@ public class Main {
     public static void main(String[] args) {
         // 여기에 코드를 작성해주세요.
         Scanner sc = new Scanner(System.in);
-        int a = sc.nextInt();
-        int b = sc.nextInt();
+        double a = sc.nextInt();
+        double b = sc.nextInt();
         System.out.printf("%.2f",(a+b)/(a-b));
     }
 }
